Update mist-guests-stats.py
Handled an issue where the GET request fails. If that happens, we don't register the guest because it was probably a quick connection and disconnection to the SSID. The Websocket will send a new message if the guest reconnects so we get another chance to send a new GET request.